Update an integral datatype; rebalance thread jobs.

+++ src/main/java/tv/lastpixel/PrimeWalker.java
@@ -37,7 +37,7 @@ public class PrimeWalker implements Runnable {
         } else if (n % 2 == 0) {
             return false;
         }
-        for (int i = 3; i * i <= n; i += 2) {
+        for (long i = 3; i * i <= n; i += 2) {
             if (n % i == 0) {
                 return false;
             }
